@@ -33,13 +33,13 @@
     "i18n-js": "http://github.com/fnando/i18n-js/archive/v3.0.0.rc8.tar.gz",
     "immutable": "~3.8.1",
     "ncp": "~2.0.0",
-    "react": "~15.3.1",
-    "react-addons-css-transition-group": "~15.3.1",
-    "react-addons-test-utils": "~15.3.2",
-    "react-dom": "~15.3.1",
+    "react": "~15.4.2",
+    "react-addons-css-transition-group": "~15.4.2",
+    "react-addons-test-utils": "~15.4.2",
+    "react-dom": "~15.4.2",
     "react-highlight": "~0.9.0",
-    "react-router": "~2.8.0",
-    "underscore.string": "~3.3.4",
+    "react-router": "~3.0.0",
+    "underscore.string": "^3.3.4",
     "xhr-mock": "git@github.com:resin-io-modules/xhr-mock.git#improvements"
   },
   "dependencies": {
